Move \appendix call to the included file
There is a well known issue with using \include and creating the table of contents

There is a good explanation of this issue here:
http://tex.stackexchange.com/questions/50036/first-chapter-appearing-before-first-line-in-table-of-contents

Basically, when using \include for other tex files, LaTeX will begin reading,
and creating .aux files in an unexpected order. By having \appendix in the same
file as the actual Appendix, rather than in the main document ensures the
ordering is correct.